backend/app/requirements.txt
fastapi
uvicorn
httpx
pydantic[email]
python-dotenv
aiosmtplib
email-validator
httpx